Split out plotting module
