+++ src/silicone/cli.py
@@ -1,4 +1,3 @@
-import itertools
 import os.path
 
 import matplotlib.pyplot as plt
@@ -6,6 +5,8 @@ import numpy as np
 import pandas as pd
 import click
 import pyam
+
+from .plotting import plot_emissions, plot_multiple_models
 
 
 @click.command(context_settings={"help_option_names": ["-h", "--help"]})
@@ -141,7 +142,9 @@ def plot_emission_correlations(
                     smooth_quant_df.to_csv(
                         os.path.join(
                             output_dir,
-                            "{}_{}_{}.csv".format(x_gas[10:], y_gas[10:], year_of_interest),
+                            "{}_{}_{}.csv".format(
+                                x_gas[10:], y_gas[10:], year_of_interest
+                            ),
                         )
                     )
 
@@ -175,35 +178,6 @@ def plot_emission_correlations(
                     output_dir, "gases_rank_correlation_{}.csv".format(year_of_interest)
                 )
             )
-
-
-def plot_emissions(seaborn_df, x_gas, y_gas, x_units, y_units):
-    colours_for_plot = "black"
-    plt.scatter(x=x_gas, y=y_gas, label=colours_for_plot, data=seaborn_df, alpha=0.5)
-    plt.xlabel("Emissions of {} ({})".format(x_gas[10:], x_units))
-    plt.ylabel("Emissions of {} ({})".format(y_gas[10:], y_units))
-
-
-def plot_multiple_models(legend_fraction, seaborn_df, x_gas, y_gas, x_units, y_units):
-    fig = plt.figure(figsize=(16, 12))
-    ax = plt.subplot(111)
-    all_models = list(seaborn_df["model"].unique())
-    markers = itertools.cycle(["s", "o", "v", "<", ">", ","])
-    for model in all_models:
-        to_plot = np.where(seaborn_df["model"] == model)[0]
-        if any(to_plot):
-            plt.scatter(
-                x=seaborn_df[x_gas].loc[to_plot],
-                y=seaborn_df[y_gas].loc[to_plot],
-                label=model,
-                alpha=0.5,
-                marker=next(markers),
-            )
-    box = ax.get_position()
-    ax.set_position([box.x0, box.y0, box.width * legend_fraction, box.height])
-    ax.legend(loc="center left", bbox_to_anchor=(1, 0.5))
-    plt.xlabel("Emissions of {} ({})".format(x_gas[10:], x_units))
-    plt.ylabel("Emissions of {} ({})".format(y_gas[10:], y_units))
 
 
 def rolling_window_find_quantiles(xs, ys, quantiles, nboxes=10, decay_length_factor=1):
